No need for CodgenTypeHelper to be namespaced under CodeGen.  Fixes circular require.
